avoid potential circular dependency
