+++ src/components/Common/PluginErrorBoundary.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+
+export class PluginErrorBoundary extends React.Component<
+  { children: React.ReactNode; pluginName: string; fallback?: React.ReactNode },
+  { hasError: boolean }
+> {
+  constructor(props: {
+    children: React.ReactNode;
+    pluginName: string;
+    fallback?: React.ReactNode;
+  }) {
+    super(props);
+    this.state = { hasError: false };
+  }
+
+  static getDerivedStateFromError() {
+    return { hasError: true };
+  }
+
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    console.error(
+      `[Plugin Error] Plugin "${this.props.pluginName}" encountered an error:`,
+      error,
+      errorInfo,
+    );
+  }
+
+  render() {
+    if (this.state.hasError) {
+      return this.props.fallback || null;
+    }
+
+    return this.props.children;
+  }
+}
